requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
fonttools==4.61.1
frozenlist==1.8.0
greenlet==3.3.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-groq==0.1.5
langchain-openai==1.1.7
langchain-text-splitters==0.0.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.1.147
marshmallow==3.26.2
matplotlib==3.10.8
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==23.2
pandas==2.3.3
pillow==12.1.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
ruff==0.14.13
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
